backend/app/services/forecasting.py
import pandas as pd
import numpy as np
from typing import Dict, List, Optional, Tuple
from statsmodels.tsa.arima.model import ARIMA
from prophet import Prophet
from lightgbm import LGBMRegressor
from xgboost import XGBRegressor
from sklearn.metrics import mean_absolute_error, mean_squared_error
from app.models.forecast import Forecast
from app.models.schema import SchemaMapping
from sqlalchemy.orm import Session

class ForecastingService:
    @staticmethod
    def select_model(data: pd.DataFrame, frequency: str) -> Tuple[str, Dict]:
        """Select the best model based on data characteristics."""
        # Basic data analysis
        n_samples = len(data)
        has_seasonality = ForecastingService._check_seasonality(data)
        has_trend = ForecastingService._check_trend(data)
        
        # Model selection logic
        if n_samples < 50:
            return "ARIMA", {"order": (1, 1, 1)}
        elif has_seasonality and has_trend:
            return "Prophet", {"seasonality_mode": "multiplicative"}
        elif n_samples > 1000:
            return "LightGBM", {"n_estimators": 100}
        else:
            return "XGBoost", {"n_estimators": 100}
    
    @staticmethod
    def _check_seasonality(data: pd.DataFrame) -> bool:
        """Check if the data shows seasonality."""
        # Implement seasonality detection logic
        return False
    
    @staticmethod
    def _check_trend(data: pd.DataFrame) -> bool:
        """Check if the data shows trend."""
        # Implement trend detection logic
        return False
    
    @staticmethod
    def generate_forecast(
        data: pd.DataFrame,
        model_name: str,
        model_params: Dict,
        horizon: int,
        frequency: str
    ) -> Tuple[pd.DataFrame, Dict]:
        """Generate forecast using the selected model."""
        if model_name == "ARIMA":
            return ForecastingService._forecast_arima(data, model_params, horizon)
        elif model_name == "Prophet":
            return ForecastingService._forecast_prophet(data, model_params, horizon)
        elif model_name == "LightGBM":
            return ForecastingService._forecast_lightgbm(data, model_params, horizon)
        elif model_name == "XGBoost":
            return ForecastingService._forecast_xgboost(data, model_params, horizon)
        else:
            raise ValueError(f"Unsupported model: {model_name}")
    
    @staticmethod
    def _forecast_arima(data: pd.DataFrame, params: Dict, horizon: int) -> Tuple[pd.DataFrame, Dict]:
        """Generate forecast using ARIMA model."""
        model = ARIMA(data, order=params["order"])
        results = model.fit()
        forecast = results.forecast(steps=horizon)
        
        metrics = {
            "mae": mean_absolute_error(data, results.fittedvalues),
            "rmse": np.sqrt(mean_squared_error(data, results.fittedvalues))
        }
        
        return forecast, metrics
    
    @staticmethod
    def _forecast_prophet(data: pd.DataFrame, params: Dict, horizon: int) -> Tuple[pd.DataFrame, Dict]:
        """Generate forecast using Prophet model."""
        model = Prophet(**params)
        model.fit(data)
        future = model.make_future_dataframe(periods=horizon)
        forecast = model.predict(future)
        
        metrics = {
            "mae": mean_absolute_error(data["y"], model.history["yhat"]),
            "rmse": np.sqrt(mean_squared_error(data["y"], model.history["yhat"]))
        }
        
        return forecast, metrics
    
    @staticmethod
    def _forecast_lightgbm(data: pd.DataFrame, params: Dict, horizon: int) -> Tuple[pd.DataFrame, Dict]:
        """Generate forecast using LightGBM model."""
        model = LGBMRegressor(**params)
        # Implement LightGBM forecasting logic
        return pd.DataFrame(), {}
    
    @staticmethod
    def _forecast_xgboost(data: pd.DataFrame, params: Dict, horizon: int) -> Tuple[pd.DataFrame, Dict]:
        """Generate forecast using XGBoost model."""
        model = XGBRegressor(**params)
        # Implement XGBoost forecasting logic
        return pd.DataFrame(), {}
    
    @staticmethod
    def create_forecast(
        db: Session,
        user_id: int,
        schema_mapping_id: int,
        name: str,
        description: Optional[str],
        forecast_horizon: int,
        forecast_frequency: str,
        start_date: pd.Timestamp,
        end_date: pd.Timestamp,
        selected_model: str,
        model_parameters: Dict,
        model_metrics: Dict,
        forecast_values: pd.DataFrame,
        confidence_intervals: pd.DataFrame
    ) -> Forecast:
        """Create a new forecast record."""
        forecast = Forecast(
            user_id=user_id,
            schema_mapping_id=schema_mapping_id,
            name=name,
            description=description,
            forecast_horizon=forecast_horizon,
            forecast_frequency=forecast_frequency,
            start_date=start_date,
            end_date=end_date,
            selected_model=selected_model,
            model_parameters=model_parameters,
            model_metrics=model_metrics,
            forecast_values=forecast_values.to_dict(),
            confidence_intervals=confidence_intervals.to_dict()
        )
        
        db.add(forecast)
        db.commit()
        db.refresh(forecast)
        return forecast
    
    @staticmethod
    def get_forecast(db: Session, forecast_id: int) -> Optional[Forecast]:
        """Get a forecast by ID."""
        return db.query(Forecast).filter(Forecast.id == forecast_id).first()
    
    @staticmethod
    def get_user_forecasts(db: Session, user_id: int) -> List[Forecast]:
        """Get all forecasts for a user."""
        return db.query(Forecast).filter(Forecast.user_id == user_id).all() 