backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
statsmodels==0.14.0
prophet==1.1.4
lightgbm==4.1.0
xgboost==2.0.2
python-multipart==0.0.6
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
